eval_code/requirements.txt
# Core dependencies
requests>=2.28.0
aiohttp>=3.8.0
pyyaml>=6.0
tqdm>=4.64.0

# Async support
asyncio-throttle>=1.0.0

# Optional: For OpenAI API
openai>=1.0.0

# Optional: For Anthropic API
anthropic>=0.7.0

# Data processing
numpy>=1.23.0

